+++ objects.c
@@ -19,6 +19,13 @@ Cursor wait_cursor;
 GC gc_text_black;
 GC gc_text_white;
 Colormap colormap;
+XColor color_new_window;
+XColor color_move_window;
+XColor color_resize_window;
+XColor color_minimize_window;
+XColor color_maximize_window;
+XColor color_close_window;
+XColor color_exit;
 
 void objects_init(void) {
     display = XOpenDisplay(NULL);
@@ -58,4 +65,19 @@ void objects_init(void) {
 	gc_text_white = XCreateGC(display, window, GCForeground | GCFont, &gcv_text_white);
 
     colormap = XDefaultColormapOfScreen(screen);
+
+    color_new_window = (XColor) { .red = 0, .green = 40000, .blue = 0 };
+	color_move_window = (XColor) { .red = 0, .green = 0, .blue = 65535 };
+	color_resize_window = (XColor) { .red = 30000, .green = 30000, .blue = 0 };
+	color_minimize_window = (XColor) { .red = 0, .green = 0, .blue = 30000 };
+	color_maximize_window = (XColor) { .red = 30000, .green = 0, .blue = 30000 };
+	color_close_window = (XColor) { .red = 65535, .green = 0, .blue = 0 };
+	color_exit = (XColor) { .red = 0, .green = 0, .blue = 0 };
+	XAllocColor(display, colormap, &color_new_window);
+	XAllocColor(display, colormap, &color_move_window);
+	XAllocColor(display, colormap, &color_resize_window);
+	XAllocColor(display, colormap, &color_minimize_window);
+	XAllocColor(display, colormap, &color_maximize_window);
+	XAllocColor(display, colormap, &color_close_window);
+	XAllocColor(display, colormap, &color_exit);
 }